Added file exists check
We added some check that are performed during the rebuild of the data base. In case a file that needs to get importet we check befor import /processing if the file exists. In case it does not exists we create a warning in the debug window.. We do not create a popup message.